ora SELF_ID, MASTER_ID E SLAVE_ID sono variabili atomiche

+++ include/uart_headers/id_type.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <cstdint>
+#include <atomic>
 
-
-typedef int8_t module_id_t; //modo comodo per cambiare il tipo di ID se necessario
+typedef int module_id_t; //modo comodo per cambiare il tipo di ID se necessario
